# internal/pkg/api/z_defaultService.gen_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.34.1
// source: z_defaultService.gen.proto

package api

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationDefaultCURDAddSheetCeshibiao = "/api.defaultCURD/AddSheetCeshibiao"
const OperationDefaultCURDDelSheetCeshibiaoByIDList = "/api.defaultCURD/DelSheetCeshibiaoByIDList"
const OperationDefaultCURDGetSheetCeshibiaoList = "/api.defaultCURD/GetSheetCeshibiaoList"
const OperationDefaultCURDUpdateSheetCeshibiao = "/api.defaultCURD/UpdateSheetCeshibiao"

type DefaultCURDHTTPServer interface {
	// AddSheetCeshibiao --------------------------------------------------
	// tbl : sheet_ceshibiao
	AddSheetCeshibiao(context.Context, *AddSheetCeshibiaoRequest) (*AddSheetCeshibiaoResponse, error)
	DelSheetCeshibiaoByIDList(context.Context, *DelSheetCeshibiaoByIDListRequest) (*Empty, error)
	GetSheetCeshibiaoList(context.Context, *GetSheetCeshibiaoListRequest) (*GetSheetCeshibiaoListResponse, error)
	UpdateSheetCeshibiao(context.Context, *UpdateSheetCeshibiaoRequest) (*UpdateSheetCeshibiaoResponse, error)
}

func RegisterDefaultCURDHTTPServer(s *http.Server, srv DefaultCURDHTTPServer) {
	r := s.Route("/")
	r.POST("/sheet-ceshibiao", _DefaultCURD_AddSheetCeshibiao0_HTTP_Handler(srv))
	r.GET("/sheet-ceshibiao", _DefaultCURD_GetSheetCeshibiaoList0_HTTP_Handler(srv))
	r.PATCH("/sheet-ceshibiao", _DefaultCURD_UpdateSheetCeshibiao0_HTTP_Handler(srv))
	r.DELETE("/sheet-ceshibiao", _DefaultCURD_DelSheetCeshibiaoByIDList0_HTTP_Handler(srv))
}

func _DefaultCURD_AddSheetCeshibiao0_HTTP_Handler(srv DefaultCURDHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AddSheetCeshibiaoRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDefaultCURDAddSheetCeshibiao)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.AddSheetCeshibiao(ctx, req.(*AddSheetCeshibiaoRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*AddSheetCeshibiaoResponse)
		return ctx.Result(200, reply)
	}
}

func _DefaultCURD_GetSheetCeshibiaoList0_HTTP_Handler(srv DefaultCURDHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetSheetCeshibiaoListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDefaultCURDGetSheetCeshibiaoList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetSheetCeshibiaoList(ctx, req.(*GetSheetCeshibiaoListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetSheetCeshibiaoListResponse)
		return ctx.Result(200, reply)
	}
}

func _DefaultCURD_UpdateSheetCeshibiao0_HTTP_Handler(srv DefaultCURDHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateSheetCeshibiaoRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDefaultCURDUpdateSheetCeshibiao)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateSheetCeshibiao(ctx, req.(*UpdateSheetCeshibiaoRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateSheetCeshibiaoResponse)
		return ctx.Result(200, reply)
	}
}

func _DefaultCURD_DelSheetCeshibiaoByIDList0_HTTP_Handler(srv DefaultCURDHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DelSheetCeshibiaoByIDListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDefaultCURDDelSheetCeshibiaoByIDList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DelSheetCeshibiaoByIDList(ctx, req.(*DelSheetCeshibiaoByIDListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Empty)
		return ctx.Result(200, reply)
	}
}

type DefaultCURDHTTPClient interface {
	// AddSheetCeshibiao --------------------------------------------------
	// tbl : sheet_ceshibiao
	AddSheetCeshibiao(ctx context.Context, req *AddSheetCeshibiaoRequest, opts ...http.CallOption) (rsp *AddSheetCeshibiaoResponse, err error)
	DelSheetCeshibiaoByIDList(ctx context.Context, req *DelSheetCeshibiaoByIDListRequest, opts ...http.CallOption) (rsp *Empty, err error)
	GetSheetCeshibiaoList(ctx context.Context, req *GetSheetCeshibiaoListRequest, opts ...http.CallOption) (rsp *GetSheetCeshibiaoListResponse, err error)
	UpdateSheetCeshibiao(ctx context.Context, req *UpdateSheetCeshibiaoRequest, opts ...http.CallOption) (rsp *UpdateSheetCeshibiaoResponse, err error)
}

type DefaultCURDHTTPClientImpl struct {
	cc *http.Client
}

func NewDefaultCURDHTTPClient(client *http.Client) DefaultCURDHTTPClient {
	return &DefaultCURDHTTPClientImpl{client}
}

// AddSheetCeshibiao --------------------------------------------------
// tbl : sheet_ceshibiao
func (c *DefaultCURDHTTPClientImpl) AddSheetCeshibiao(ctx context.Context, in *AddSheetCeshibiaoRequest, opts ...http.CallOption) (*AddSheetCeshibiaoResponse, error) {
	var out AddSheetCeshibiaoResponse
	pattern := "/sheet-ceshibiao"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationDefaultCURDAddSheetCeshibiao))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *DefaultCURDHTTPClientImpl) DelSheetCeshibiaoByIDList(ctx context.Context, in *DelSheetCeshibiaoByIDListRequest, opts ...http.CallOption) (*Empty, error) {
	var out Empty
	pattern := "/sheet-ceshibiao"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationDefaultCURDDelSheetCeshibiaoByIDList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *DefaultCURDHTTPClientImpl) GetSheetCeshibiaoList(ctx context.Context, in *GetSheetCeshibiaoListRequest, opts ...http.CallOption) (*GetSheetCeshibiaoListResponse, error) {
	var out GetSheetCeshibiaoListResponse
	pattern := "/sheet-ceshibiao"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationDefaultCURDGetSheetCeshibiaoList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *DefaultCURDHTTPClientImpl) UpdateSheetCeshibiao(ctx context.Context, in *UpdateSheetCeshibiaoRequest, opts ...http.CallOption) (*UpdateSheetCeshibiaoResponse, error) {
	var out UpdateSheetCeshibiaoResponse
	pattern := "/sheet-ceshibiao"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationDefaultCURDUpdateSheetCeshibiao))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}