# openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /course-swap-request:
        get:
            tags:
                - SchoolCURD
            operationId: SchoolCURD_GetCourseSwapRequestList
            parameters:
                - name: IDList
                  in: query
                  schema:
                    type: array
                    items:
                        type: integer
                        format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.GetCourseSwapRequestListResponse'
        post:
            tags:
                - SchoolCURD
            description: |-
                --------------------------------------------------
                 tbl : course_swap_request
            operationId: SchoolCURD_AddCourseSwapRequest
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.AddCourseSwapRequestRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.AddCourseSwapRequestResponse'
        delete:
            tags:
                - SchoolCURD
            operationId: SchoolCURD_DelCourseSwapRequestByIDList
            parameters:
                - name: IDList
                  in: query
                  schema:
                    type: array
                    items:
                        type: integer
                        format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.Empty'
        patch:
            tags:
                - SchoolCURD
            operationId: SchoolCURD_UpdateCourseSwapRequest
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.UpdateCourseSwapRequestRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.UpdateCourseSwapRequestResponse'
...
    /sheet-ceshibiao:
        get:
            tags:
                - DefaultCURD
            operationId: DefaultCURD_GetSheetCeshibiaoList
            parameters:
                - name: IDList
                  in: query
                  schema:
                    type: array
                    items:
                        type: integer
                        format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.GetSheetCeshibiaoListResponse'
        post:
            tags:
                - DefaultCURD
            description: |-
                --------------------------------------------------
                 tbl : sheet_ceshibiao
            operationId: DefaultCURD_AddSheetCeshibiao
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.AddSheetCeshibiaoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.AddSheetCeshibiaoResponse'
        delete:
            tags:
                - DefaultCURD
            operationId: DefaultCURD_DelSheetCeshibiaoByIDList
            parameters:
                - name: IDList
                  in: query
                  schema:
                    type: array
                    items:
                        type: integer
                        format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.Empty'
        patch:
            tags:
                - DefaultCURD
            operationId: DefaultCURD_UpdateSheetCeshibiao
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.UpdateSheetCeshibiaoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.UpdateSheetCeshibiaoResponse'
...
components:
    schemas:
...
        api.AddCourseSwapRequestRequest:
            type: object
            properties:
                courseSwapRequest:
                    $ref: '#/components/schemas/api.CourseSwapRequestInfo'
        api.AddCourseSwapRequestResponse:
            type: object
            properties:
                courseSwapRequest:
                    $ref: '#/components/schemas/api.CourseSwapRequestInfo'
...
        api.AddSheetCeshibiaoRequest:
            type: object
            properties:
                sheetCeshibiao:
                    $ref: '#/components/schemas/api.SheetCeshibiaoInfo'
        api.AddSheetCeshibiaoResponse:
            type: object
            properties:
                sheetCeshibiao:
                    $ref: '#/components/schemas/api.SheetCeshibiaoInfo'
...
        api.CourseSwapRequestInfo:
            type: object
            properties:
                ID:
                    type: integer
                    format: int32
                srcTeacher:
                    type: string
                srcDate:
                    type: string
                srcCourseNum:
                    type: integer
                    format: int32
                srcCourse:
                    type: string
                srcClass:
                    type: string
                dstTeacher:
                    type: string
                dstDate:
                    type: string
                dstCourseNum:
                    type: integer
                    format: int32
                dstCourse:
                    type: string
                dstClass:
                    type: string
                createTime:
                    type: string
                status:
                    type: integer
                    format: int32
            description: |-
                --------------------------------------------------
                 tbl : course_swap_request
...
        api.Empty:
            type: object
            properties: {}
...
        api.GetCourseSwapRequestListResponse:
            type: object
            properties:
                courseSwapRequestList:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.CourseSwapRequestInfo'
...
        api.GetSheetCeshibiaoListResponse:
            type: object
            properties:
                sheetCeshibiaoList:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.SheetCeshibiaoInfo'
...
        api.SheetCeshibiaoInfo:
            type: object
            properties:
                ID:
                    type: integer
                    format: int32
                lastEditFrom:
                    type: string
                jingtou:
                    type: string
                changci:
                    type: string
                zhuangtai:
                    type: string
            description: |-
                --------------------------------------------------
                 tbl : sheet_ceshibiao
...
        api.UpdateCourseSwapRequestRequest:
            type: object
            properties:
                courseSwapRequest:
                    $ref: '#/components/schemas/api.CourseSwapRequestInfo'
        api.UpdateCourseSwapRequestResponse:
            type: object
            properties:
                courseSwapRequest:
                    $ref: '#/components/schemas/api.CourseSwapRequestInfo'
...
        api.UpdateSheetCeshibiaoRequest:
            type: object
            properties:
                sheetCeshibiao:
                    $ref: '#/components/schemas/api.SheetCeshibiaoInfo'
        api.UpdateSheetCeshibiaoResponse:
            type: object
            properties:
                sheetCeshibiao:
                    $ref: '#/components/schemas/api.SheetCeshibiaoInfo'
...
servers:
    - description: PGO API
      url: http://127.0.0.1:8080